Enhanced fsi-wserver and fsi-web so they no longer have the opportunity to pass user input to a shell command.  Previously, both of these functions passed a url as an argument to w3m.  The new versions use url-retrieve to download the contents without calling out to the shell. url-retrieve returns both the headers and body (separated by a blank line) into a buffer.
fsi-wserver now pulls the headers from this buffer and sends the output
to the user.  fsi-web has a little more work to do because the body is
raw HTML.  I use shell-command-on-region to pipe the body into w3m (not
passed as an argument) and then have the function replace the contents
of the region with the now formatted output.  This is then returned to
the user.